More resbar fixery.
Reduced Katyusha damage to 1800 (from 4000), increased wobble to 4000 (from 3000), reduced reload cost to 400 (from 560); according to Wikipedia RS-132 only carried 0.9 kg of explosive, but this is disputed elsewhere? In any case, the net effect will be that the Katyusha is much less effective against vehicles and buildings, but costs less to fire.

Increased M-30 damage to 7200 (from 5000).