Return AEC to TD-only role.
It was causing problems as a "fast light tank":

1) It was quite good at a wide range of roles (inf, vehs, raiding, anti armor
-- nyanryan noted that it penetrates T-34 at max range).
2) It made the TD yard upgrade for brits too safe. You get a great TD, and
a great pseudo-tank.

The alternative to this mildly ahistorical tweak is to just remove the AEC
from the brit TD yard upgrade -- that would help, but it would remain a pretty
problematic "sort of good at everything" unit. This change narrows the role
considerably: AEC becomes a "brawler TD." It has been priced accordingly.