Don't count cloaking AT guns and SmallTankShelter
Here are the remaining units that cloak but are not 'dontCount'.

01:43:53 [spring1944.sdd] bt bt-desk$ git grep -l cloakCost | xargs grep -L dontCount
units/GMToolbox.fbi
units/rus/buildings/RUSPShack.fbi
units/rus/infantry/RUSCommissar.fbi

Of those, I'm pretty sure PShack isn't in the build tree, so we should nuke it.
GMToolbox and Commissar are good (although the potential for a hiding commissar
to muck up the game is possible, presumably you'll be found when you try to use
him to rebuild).