Rethink partisans a bit.
Every spawn point now drains -5. This may seem like a
lot (it is), but they've been beefed up accordingly:

1) partisans now cap like regular soviet troops
2) partisans now use the same rifle as regular sov riflemen
3) they now spawn at 5 second intervals (i.e. crazy fast).
4) the drop cost has been lowered to 500

The only added restriction is that they will cease spawning if an enemy
gets too close -- that is, within their supply range.

The hope here is that partisans become a more intentional tool.

I'm reasonably sure that they're still no match for regular infantry head to
head (with the interesting property that the more you retreat from the
partisans, the more expensive they are for the other player), but the ability
to consistently pressure somewhere with rapid reinforcement is pretty powerful.
If this turns out to be too much, we can turn down the spawn frequency, raise
the drain, or add something else like causing the spawn point to degrade if the
player doesn't have enough money to run it.