Remove side_armor_defs in favor of damageGroup.
This adds (hopefully appropriate) damageGroup tags to all of the unit base
classes, which frees us from maintaining a list of all the units in the game
in various side_armor_def/<side>.lua lists.

Aside from that, this commit introduces the 'Bunker' interface (which puts
units in the 'bunkers' damageGroup) and changed Halftrack to inherit from
ArmouredCar, so they share the armouredVehicles damageGroup.

Finally, this commit renames the 'flag' damage group to 'invincible', since it
includes more than flags these days. This might be the wrong choice for
squad_spawn units and the like, but we'll see.