s/Echo/Log/g cleanup: gamedata/
This is the first commit in a series which swap Spring.Echo for Spring.Log.
This is a good thing because it means players don't see technical printout
details that they don't need -- it just goes to the infolog.

This commit also removes a number of commented out Echo usages.