Standardize mortar squad cost: 2k + single unit HP
This is a reasonable change from JAL's mutator: mortar capabilities are pretty
much the same between sides (the brits do indeed have a good mortar, but
probably not detectably better, and almost certainly not 900 command better
than the soviet one). The small difference for side HP pretty much reflects the
tiny difference in 'time to save the unit when it starts getting shot at'.