Bandaid #150
This is a partial revert of ac6b2c3d047a505367e131862e45a9f49f6abdb6, which
exposed (but did not cause) this issue.

The root problem is that paralysis (as inflicted by the morph gadget while
morphing) resets the reload state. As a result, the reload frame-based 'did the
weapon fire' detection assumes that the weapon did indeed fire, and incurs ammo
usage.

ac6b2c3d047a505367e131862e45a9f49f6abdb6 exposed this issue by running
ProcessWeapon every frame, and subtracting multiple rounds of ammo per
ammo-using weapon that seemed to have fired. In the case here, that was all the
ammo using weapons, so it could subtract as many as 2 rounds per frame for
a few frames.

The correct fix for this problem is to eliminate any detection of weapon firing
from game_ammo.lua and move it into the unit scripts.