Hopefully fix crash on mine laying
Game used to crash when multiple engineers (like 10) laid
mines on the same spot. Issue is probably that game tries to
place every mine on different spot and it gets stuck trying to do
that as space is not enough. Changed code so that it stops
trying after mine has hit already used spot three times. Not
optimal solution either but at least it avoids crash on my
machine.