Fix Flag.lua crash from errant faction picker messages
The cause was faction picker being restarted during the game and
re-sending the luaRulesMsg about setting the faction. The game_setup.lua
recvLuaRulesMsg callback was active throughout the game, and as a result the
player's side in GG.teamSide was reset to the content of the message, which was
"random team (GM)".

Fixes:
    * faction_picker no longer sends the message in Initialize if the game is
    started
    * game_setup.lua bails out if gameframe > 0 in the message handler
    * game_setup removes itself from the list of gadgets after doing its work
    (spawning bases)

Any one of these would have been sufficient, but this is an annoying bug to
track, so defense in depth++.