GERSBoot: Remove maxAngleDiff from dummy weapon
and tweak the turret firing angles so they can both fire

This stops the unit from turning to face its target on an attack order, and
makes the front turret able to fire in a much less restricted range (previously
it could only fire if it were facing the target dead-on, which doesn't make
much sense for a turret).

Slightly reduce rear turret firing arc to match the front. This provides a good
set of overlap for the turrets when the ship is broadside, but without giving
it double-fire from too many forward angles.